Pass a reference to a Box'ed unboxed closure, in Timer.
So that it goes like this:

   &'a      Box<FnMut() -> uint + 'a>
   ^        ^
   1 word   2 words (vtable and pointer)

We can safely cast the reference to a void*, and at the same time we
know that the reference is going to be valid as long as the closure is.

Before, I had only the Box and took a reference to it when the time to
transmute it into void* came.  However, this was broken because the
reference that we transmuted wasn't valid after Timer::start() finished.

This is very ugly but I'm not sure how to do it differently.