Add AudioCVT Back, use a Rust-allocated Vec<u8>.
This uses a rust-allocated vector; the vector is grown in-place and
the interior pointer is passed to SDL.

Upon success: the vector's actual length is updated and the vector
is returned to the caller.

In the Error case; since it is an ordinary Rust Vec<T> which we
have ownership of: the memory should be deallocated as appropriate.

This incl. a test-case which converts from mono to stereo.