Changes to compile under Rust 1.0.0-alpha.
I have no idea what I'm doing. I took a lot of these changes from the PR
in rust-sdl2 that updated it for 1.0:

https://github.com/AngryLawyer/rust-sdl2/pull/272

This library has a few additional things that weren't covered there, so
I did some Googling. I am *very* new to Rust and don't know much, and so
I am not at all confident that all of these changes are right.
Particularly the bit that changes &[u8] to &str. I'm sure that's not
correct, but it does compile and the example app runs properly on my
machine using the 1.0.0-alpha rustc build.