Avoid heap allocation in AudioCVT, convert() can never fail
The size of SDL_AudioCVT is known at compile-time, so no heap
allocation is necessary.

convert() can only fail if buf is NULL, which, in this case, it never
can be. Additionally, the function can still succeed if no conversion
is needed; just do nothing to the buffer!

AudioCVT.len unfortunately uses an int and not a size_t,
so an overflow check is added to make sure the buffer length
is within c_int.