Update `Timer` API for unboxed closures
The timer now takes a `Box<F>` where F implements a closure w/ the
signature `Fn() -> u32+Send`

The timer starts as soon as it is created, and ends under the following
conditions:

* The callback, when invoked, returns a non-positive continuation interval
* The user drops the `Timer`, either by calling `drop(timer)` or by
  letting it go out of scope.

The timer is cancelled when exiting scope to ensure memory-safety. Since
the timer owns the closure's environment: the closure cannot safely be
called once the timer itself has been freed.

---

The `Send` bound implies that closures must be able to move to another
thread. This is because SDL does not guarantee that timers will be run
on the current thread.

This means mutable access to data will have to be done in a thread-safe
manner using the primitives in `std::sync`.

---

* The tests have been updated to reflect the new API.

* They have also been made more robust to intermittent failure. The
offending test now has an upper limit, after which it will cancel the
timer.

* This is meant to ensure the mutex will be available for locking
when the test checks the exit condition.