Update, simplify audio API
* Move `AudioSpecDesired::open_audio_device` to `AudioDevice`
* Remove superfluous AudioCallbackUserdata
* Initialize the audio callback after the obtained AudioSpec is known

Just simplifying the audio API so that it's not the "Audio specification"
that's opening the audio device. I think this is less confusing.

SDL2 doesn't support recording devices yet, so the feature was
removed.
The feature can be added back in when SDL2 does support recording.

Initializing the audio callback once the AudioSpec is known allows the callback
to be initialized with a known frequency, channel count and sample size.