Window safety: Ensure thread, event-pump safety
All Window properties must be accessed using `Window::properties`,
which requires a reference to the event pump.
If the event pump is not being used (mutably), then the caller can
obtain a shared reference to it, which means the event pump cannot
mutate the Window.
Renderer also has a way to access window properties, using
`Renderer::window::properties`.

Additionally, Window initialization requires an `Sdl` reference.
Because `Sdl` is `!Send` and `!Sync`, obtaining a `Sdl` reference is
only possible on the same thread.