Make `SurfaceRef` 0 bytes
In PR #435, I made `SDL_Surface` a DST. This is actually invalid
because the resulting pointer `*mut SDL_Surface` became a fat pointer
(i.e. 16 bytes instead of 8 on 64-bit targets).
Unsized types (as opposed to dynamically sized types) in Rust aren't
currently possible.

To prevent partial swapping and memory corruption with
`std::mem::swap()`, the structure is 0 bytes.
This shouldn't cause any problems, as `SurfaceRef` is always a
reference to SDL-owned memory.