Implement tests for registering custom events.
The change in src/sdl2/lib.rs to export the libc crate is necessary to
make sure that the downstream user can cast the data1 and data2 pointers
from/to custom pointers.

Also, the test harness is disabled for the tests/events.rs file, since
SDL can only be initialized once and doesn't support multithreading.