Convert SDL_image bindings into autogenerated ones
I'm not sure if this is the intended way of using bindgen, but the
SDL_image bindings now piggy-back on the core SDL bindings by
blacklisting some types that gets dragged into the image module.
The solution probably does not scale, but it might be preferable to have
autoupdateable bindings rather than having to do it manually.

Progresses on #647