Remove cfg!(target_os="x") from build.rs
using the cfg macro breaks cross compilation. This is because the cfg
macro gets evaluated during compile time of the binary getting built. In
the case of build.rs, it gets built into it's own binary targeting the
local machine. (Which then gets run to evaluate build rules for the
primary project binary). So somewhat confusingly, the cfg(target_os..)
evaluates differently in build.rs than the rest of the project in the
case of cross compilation. Particularly for SDL, this is important for
targeting emscripten. The current build.rs breaks when trying to compile
for the web.