Fix build on Linux to support sndio dynamic loading
Current SDL2 versions have a bug: somebody forgot to declare
SNDIO_SHARED in CMakeLists.txt so even if the support for dynamically
loading sndio is fully implemented (like in all other audio backends),
it is never activated.

This in turns means that supporting sndio always requires linking at
link-time against libsndio, which in turns creates binaries that require
libsndio on the target system to even start (even if no audio support
is required, or if other audio options like PulseAudio or Jack are
available).

build.rs had a workaround for this bug: it always linked against
libsndio, but this created binaries with this dependency.

I've opened a bug upstream:
https://bugzilla.libsdl.org/show_bug.cgi?id=5105

and I'm including the one-line patch in this PR so that we can fix
rust-sdl2 immediately.