audio: fix device name use-after-free in AudioDevice::open
Fixes a use-after-free in `AudioDevice::open`, which occurs when
selecting a particular device by name (as opposed to using a default
device).

Specifically, when extracting a C-style string pointer from a
`Option<CString>`, the option was consumed, its contents turned into a
pointer, and the original `CString` dropped. After that the C-style
string pointer is dangling, as its backing rust CString has been freed.

To fix this, the CString must be kept alive, which is achieved by simply
creating an intermediary `Option<&CString>`, and consuming that.