sdl2-sys: Allow bundled debug build
sdl2-sys was hardcoding the bundled build as a release build,
unconditionally for a long time.  This can cause problems however if an
upstream link of a binary is mixing C++ libraries that mix debug and non
debug builds, or so at least I'm led to believe for static builds.