rust-sdl2: cherry pick vs2019 intrisics fix
When building for "bundled" dynamic library on Windows, VS2019
encounters a link error as it is generating an intrinsic for memset but
the cmake build misses linking in the code for it.

Fix by cherrypicking upstream patch that works around the problem by
telling the compiler to generate actual function calls to it.

https://github.com/libsdl-org/SDL/issues/3865