Bounce through SDL2 heap in AudioCVT::convert
This commit rewrites AudioCVT::convert to bounce the audio buffer into a
an SDL2 heap allocation, rather than trying to reuse the rust heap
buffer.  This is critical, as the underlying library warns internally
that the buffer may be reallocated, breaking configurations where
the library is not using the same heap as rust.

The underlying implementation also notes that the underlying buffer may
be transparently resized to larger than the output as part of the
transformations, so relying on the buffer being capacity() bytes long to
ensure the buffer is not re-allocated is a broken assumption.

Closes #1096