Update doc for AudioCVT::convert
Update to reflect the surface interface and remove the bits about
conversion in place which are now out of date.

This also removes the note about buffer overflows in issue #270. Looking
back, the understanding was that the buffer wouldn't get reallocated,
and this is evident even in the sample C code used there.