Copy Linux dynamic libs and symlinks to output
Just like Windows, we should be copying the dynamic library to the
target directory and deps directory.  Otherwise, execution may fail,
or, more likely, silently pick up the system libSDL2 and not the
"bundled" one as asked.

This is evident when trying to run tests on a system with the libsdl2
package uninstalled, where all the example tests fail unless they pick
up the system .so.