Update vcpkg rev to unblock Windows
Something happened that broke the `vcpkg.exe install` invocation for
sdl2 on Windows when using github actions with `windows-latest`.

I have not been able to determine what is actually causing the break,
however updating to a more recent version of vcpkg seems to have fixed
the issue, so that is what is proposed here.

The commit selected is current top of tree.

Original report: https://github.com/mcgoo/cargo-vcpkg/issues/9