Update README.md to merge "Bundled" section Codeblocks
Before this commit, there were separate instructions for build.rs in the section for the feature `bundled` on macos and linux but now there is a single merged build script that uses rust's conditional compilation features to merge the 2 codeblocks into 1 universal codeblock. None of the two codeblocks run if the operating system is windows or other.