Add missing CoreHaptics framework
This framework is used by the SDL_JOYSTICK module, but it is missing from Darwin builds.
I have confirmed that this fixes my issues on my MacOS build. I have added the directive for the iOS builds, too, but that will need to be double-checked as I don't have a workflow for that.